adding page with compliation of all drawings, changed website layout and folder structure
