--- src/Images.js
@@ -1,123 +1,0 @@
-const images = [
-  {
-    day: "Day 2",
-    reference: "https://i.imgur.com/Dn5fEEH.jpeg",
-    link: "/images/small2.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/r9ncz7/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 24",
-    reference: "https://www.youtube.com/watch?v=K8tQEXbDkWs",
-    link: "/images/small24.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/rpzlm7/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 26",
-    reference: "https://i.imgur.com/r5XV1yV.jpeg",
-    link: "/images/small26.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/rrlks4/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 29",
-    reference:
-      "https://i.guim.co.uk/img/media/8a840f693b91fe67d42555b24c6334e9298f4680/0_1476_2429_1456/master/2429.jpg",
-    link: "/images/small29.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/rtw04a/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 33",
-    reference:
-      "https://www.artnews.com/wp-content/uploads/2021/06/Screen-Shot-2021-06-07-at-10.16.45-AM-e1623087850826.png",
-    link: "/images/small33.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/rwyyny/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 36",
-    reference:
-      "https://media.npr.org/assets/img/2018/05/22/figurea-c5805941ebcd5b9ea00ed6901e0a16d1a884e77e-s800-c85.webp",
-    link: "/images/small36.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/rzdxdj/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-
-  {
-    day: "Day 50",
-    reference: "https://eturbonews.com/wp-content/uploads/2021/06/kingkong.jpg",
-    link: "/images/small50.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/sacx9w/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 51",
-    reference:
-      "https://www.rollingstone.com/wp-content/uploads/2021/01/bernie-meme-photog.jpg",
-    link: "/images/small51.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/sb134z/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 67",
-    reference:
-      "http://justfunfacts.com/wp-content/uploads/2016/09/the-thinker.jpg",
-    link: "/images/small67.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/snxcux/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 73",
-    reference: "https://www.youtube.com/watch?v=ok6pLKBD-XQ",
-    link: "/images/small73.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/ssm77g/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 75",
-    reference: "https://m.media-amazon.com/images/I/81tMB+4FPZL.jpg",
-    link: "/images/small75.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/su8h1i/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 83",
-    reference: "http://www.nrm.org/MT/graphics/TripleSelf.jpg",
-    link: "/images/small83.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/t0oprv/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 84",
-    reference: "https://youtu.be/3JXcqzJjHf0?t=17",
-    link: "/images/small84.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/t1hpdb/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 89",
-    reference: "https://www.youtube.com/watch?v=lBbQg8r8tdc&loop=0",
-    link: "/images/small89.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/t5dr8q/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 94",
-    reference: "https://m.media-amazon.com/images/I/51fUD1buqxL._AC_SY679_.jpg",
-    link: "/images/small94.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/t916nx/drawing_mike_trout_every_day_until_the_lockout_is/",
-  },
-  {
-    day: "Day 97",
-    reference:
-      "https://www.brightview.com/sites/default/files/inline-images/Field%20of%20Dreams.jpg",
-    link: "/images/small97.jpg",
-    directLink:
-      "https://old.reddit.com/r/baseball/comments/tb7r5k/drawing_mike_trout_for_the_last_time_because_the/",
-  },
-];
-
-export default images;
