--- src/components/About.css
@@ -1,3 +1,0 @@
-.new-line {
-  white-space: pre-line;
-}
