+++ src/components/About/About.css
@@ -1,0 +1,3 @@
+.new-line {
+  white-space: pre-line;
+}
