+++ src/components/Modal/Modal.css
@@ -1,0 +1,21 @@
+.Modal-style {
+  position: fixed;
+  gap: 10px;
+  padding: 10px;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: #e23737;
+  border-radius: 5px;
+  z-index: 1000;
+}
+/* padding: 50px */
+.Overlay-styles {
+  position: fixed;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-color: rgba(0, 0, 0, 0.7);
+  z-index: 1000;
+}
